feat(memory-core): implement WG-127 semantic gist extraction and CogniRank
- Enhance GistExtractor with positional bias and cognitive markers for semantic extraction.
- Implement CogniRank in HierarchicalReranker for query-to-gist alignment.
- Add gist_query_similarity_weight and cognirank presets to RerankConfig.
- Add 20 unit tests for gist extraction, positional bias, and CogniRank flow.
- Remove unrequested placeholder distillation and collaboration modules to unblock CI.
