+++ .github/workflows/ai-slop-learn.yml
@@ -193,7 +193,7 @@ jobs:
             core.info('Patterns file updated.');
 
       - name: Create Pull Request for updated AI slop patterns
-        uses: peter-evans/create-pull-request@v8
+        uses: peter-evans/create-pull-request@v6
         with:
           commit-message: "chore(ai-slop): learn from issue #${{ github.event.issue.number }} (${{ github.event.label.name }}) [skip ci]"
           title: "chore(ai-slop): learn from issue #${{ github.event.issue.number }}"
