+++ memory-cli/src/commands/episode/relationships/mod.rs
@@ -57,7 +57,6 @@ pub async fn add_relationship(
         reason,
         created_by,
         priority,
-        weight: None,
         custom_fields: std::collections::HashMap::new(),
     };
