+++ memory-cli/src/commands/relationships/core.rs
@@ -71,7 +71,6 @@ pub async fn add_relationship(
         reason,
         created_by,
         priority,
-        weight: None,
         custom_fields,
     };
