+++ memory-core/src/episode/graph_algorithms/mod.rs
@@ -14,8 +14,8 @@ use super::EpisodeRelationship;
 use super::relationship_errors::GraphError;
 
 pub use self::traversal::{
-    WeightedPathState, find_cycles_helper, find_path_dfs_helper, find_weighted_path_helper,
-    get_weighted_neighbors, has_cycle_helper, has_path_dfs_helper, topological_sort_helper,
+    find_cycles_helper, find_path_dfs_helper, has_cycle_helper, has_path_dfs_helper,
+    topological_sort_helper,
 };
 
 /// Check if a path exists from start to end using DFS.
@@ -419,42 +419,4 @@ where
     )?;
 
     Ok(cycles)
-}
-
-/// Find the path from start to end with the maximum total weight.
-///
-/// This is a simplified version of a shortest-path algorithm for weighted DAGs.
-///
-/// # Arguments
-///
-/// * `adjacency_list` - The graph represented as an adjacency list
-/// * `start` - The starting node (episode ID)
-/// * `end` - The target node (episode ID)
-///
-/// # Returns
-///
-/// `Ok(Option<(Vec<Uuid>, f32)>)` containing the path and its total weight,
-/// or `None` if no path exists.
-pub fn find_weighted_path<S>(
-    adjacency_list: &HashMap<Uuid, Vec<EpisodeRelationship>, S>,
-    start: Uuid,
-    end: Uuid,
-) -> Result<Option<(Vec<Uuid>, f32)>, GraphError>
-where
-    S: std::hash::BuildHasher,
-{
-    let mut state = self::traversal::WeightedPathState {
-        visited: HashSet::new(),
-        path: Vec::new(),
-        best_path: None,
-        best_weight: f32::MIN,
-    };
-
-    find_weighted_path_helper(adjacency_list, start, end, &mut state, 0.0)?;
-
-    if let Some(path) = state.best_path {
-        Ok(Some((path, state.best_weight)))
-    } else {
-        Ok(None)
-    }
 }
