+++ memory-core/src/episode/graph_algorithms/traversal.rs
@@ -129,84 +129,6 @@ where
     Ok(())
 }
 
-/// State for weighted path traversal.
-pub struct WeightedPathState {
-    pub visited: HashSet<Uuid>,
-    pub path: Vec<Uuid>,
-    pub best_path: Option<Vec<Uuid>>,
-    pub best_weight: f32,
-}
-
-/// Helper for finding the maximum weight path.
-pub fn find_weighted_path_helper<S>(
-    adjacency_list: &HashMap<Uuid, Vec<EpisodeRelationship>, S>,
-    current: Uuid,
-    target: Uuid,
-    state: &mut WeightedPathState,
-    current_weight: f32,
-) -> Result<(), GraphError>
-where
-    S: std::hash::BuildHasher,
-{
-    state.path.push(current);
-
-    if current == target {
-        if current_weight > state.best_weight {
-            state.best_weight = current_weight;
-            state.best_path = Some(state.path.clone());
-        }
-        state.path.pop();
-        return Ok(());
-    }
-
-    state.visited.insert(current);
-
-    if let Some(neighbors) = adjacency_list.get(&current) {
-        for rel in neighbors {
-            let neighbor = rel.to_episode_id;
-            let weight = rel.metadata.weight.unwrap_or(1.0);
-
-            if !state.visited.contains(&neighbor) {
-                find_weighted_path_helper(
-                    adjacency_list,
-                    neighbor,
-                    target,
-                    state,
-                    current_weight + weight,
-                )?;
-            }
-        }
-    }
-
-    state.visited.remove(&current);
-    state.path.pop();
-    Ok(())
-}
-
-/// Get weighted neighbors from an adjacency list.
-pub fn get_weighted_neighbors<S>(
-    adjacency_list: &HashMap<Uuid, Vec<EpisodeRelationship>, S>,
-    episode_id: Uuid,
-) -> Vec<(Uuid, f32, bool)>
-where
-    S: std::hash::BuildHasher,
-{
-    adjacency_list
-        .get(&episode_id)
-        .map(|rels| {
-            rels.iter()
-                .map(|r| {
-                    (
-                        r.to_episode_id,
-                        r.metadata.weight.unwrap_or(1.0),
-                        false, // Default to not being a pattern in this generic helper
-                    )
-                })
-                .collect()
-        })
-        .unwrap_or_default()
-}
-
 /// Helper for finding cycles from a specific node.
 pub fn find_cycles_helper<S>(
     adjacency_list: &HashMap<Uuid, Vec<EpisodeRelationship>, S>,
