+++ memory-core/src/retrieval/gist/config.rs
@@ -1,26 +1,26 @@
-//! Configuration for hierarchical reranking.
+//! Configuration for gist reranking.
 
 use serde::{Deserialize, Serialize};
 
-/// Configuration for hierarchical reranking.
+/// Configuration for hierarchical gist reranking.
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct RerankConfig {
     /// Weight for original relevance score (0.0-1.0)
     pub relevance_weight: f32,
     /// Weight for gist density score (0.0-1.0)
     pub density_weight: f32,
-    /// Weight for gist-to-query similarity (0.0-1.0)
+    /// Weight for gist-to-query similarity (0.0-1.0, CogniRank)
     pub gist_query_similarity_weight: f32,
     /// Weight for recency score (0.0-1.0)
     pub recency_weight: f32,
-    /// Lambda for diversity (MMR-style) (0.0-1.0)
-    /// Higher lambda = more relevance, lower = more diversity
+    /// Lambda for MMR diversity (0.0-1.0)
+    /// 1.0 = relevance only, 0.0 = diversity only
     pub diversity_lambda: f32,
     /// Maximum key points to extract per episode
     pub max_key_points: usize,
-    /// Minimum gist density threshold for inclusion
+    /// Minimum density threshold to include in results
     pub min_density_threshold: f32,
-    /// Half-life in days for recency decay
+    /// Half-life for recency decay in days
     pub recency_half_life_days: f32,
 }
 
@@ -41,15 +41,12 @@ impl Default for RerankConfig {
 
 impl RerankConfig {
     /// Create a configuration optimized for dense context.
-    ///
-    /// Prioritizes gist density and diversity for maximum information
-    /// per token in downstream prompts.
     #[must_use]
     pub fn dense() -> Self {
         Self {
             relevance_weight: 0.2,
-            density_weight: 0.4,
-            gist_query_similarity_weight: 0.25,
+            density_weight: 0.5,
+            gist_query_similarity_weight: 0.15,
             recency_weight: 0.15,
             diversity_lambda: 0.6,
             max_key_points: 2,
@@ -58,34 +55,15 @@ impl RerankConfig {
         }
     }
 
-    /// Create a configuration optimized for comprehensive context.
-    ///
-    /// Larger result set with lower density thresholds.
-    #[must_use]
-    pub fn comprehensive() -> Self {
-        Self {
-            relevance_weight: 0.35,
-            density_weight: 0.25,
-            gist_query_similarity_weight: 0.15,
-            recency_weight: 0.25,
-            diversity_lambda: 0.75,
-            max_key_points: 3,
-            min_density_threshold: 0.2,
-            recency_half_life_days: 60.0,
-        }
-    }
-
-    /// Create a configuration optimized for CogniRank.
-    ///
-    /// Prioritizes alignment between the gist and the user query.
+    /// Create a configuration optimized for CogniRank (query alignment).
     #[must_use]
     pub fn cognirank() -> Self {
         Self {
             relevance_weight: 0.2,
-            density_weight: 0.2,
-            gist_query_similarity_weight: 0.45,
-            recency_weight: 0.15,
-            diversity_lambda: 0.65,
+            density_weight: 0.3,
+            gist_query_similarity_weight: 0.4,
+            recency_weight: 0.1,
+            diversity_lambda: 0.7,
             max_key_points: 3,
             min_density_threshold: 0.3,
             recency_half_life_days: 30.0,
@@ -93,10 +71,6 @@ impl RerankConfig {
     }
 
     /// Validate the configuration.
-    ///
-    /// # Returns
-    ///
-    /// `Ok(())` if configuration is valid, `Err` with message if invalid
     pub fn validate(&self) -> Result<(), String> {
         let weight_sum = self.relevance_weight
             + self.density_weight
@@ -105,18 +79,9 @@ impl RerankConfig {
         if (weight_sum - 1.0).abs() > 0.15 {
             return Err(format!("Weights should sum to ~1.0, got {weight_sum:.2}"));
         }
-
         if !(0.0..=1.0).contains(&self.diversity_lambda) {
-            return Err(format!(
-                "diversity_lambda must be in [0.0, 1.0], got {}",
-                self.diversity_lambda
-            ));
+            return Err("diversity_lambda must be in [0.0, 1.0]".to_string());
         }
-
-        if self.max_key_points == 0 {
-            return Err("max_key_points must be at least 1".to_string());
-        }
-
         Ok(())
     }
 }
