+++ tests/e2e/mcp_relationship_chain.rs
@@ -118,7 +118,6 @@ async fn test_mcp_relationship_full_chain() {
     let metadata = RelationshipMetadata {
         reason: Some("Design leads to implementation".to_string()),
         priority: Some(9),
-        weight: None,
         ..Default::default()
     };
 
@@ -445,7 +444,6 @@ async fn test_mcp_relationship_metadata() {
     let metadata = RelationshipMetadata {
         reason: Some("Important dependency relationship".to_string()),
         priority: Some(10),
-        weight: None,
         ..Default::default()
     };
